% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{cancel_export_task}
\alias{cancel_export_task}
\title{Cancels the specified export task}
\usage{
cancel_export_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the export task.}
}
\description{
Cancels the specified export task.
}
\details{
The task must be in the \code{PENDING} or \code{RUNNING} state.
}
\section{Accepted Parameters}{
\preformatted{cancel_export_task(
  taskId = "string"
)
}
}

