% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{create_log_stream}
\alias{create_log_stream}
\title{Creates a log stream for the specified log group}
\usage{
create_log_stream(
  logGroupName = "string",
  logStreamName = "string"
)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{logStreamName}{[required] The name of the log stream.}
}
\description{
Creates a log stream for the specified log group.
}
\details{
There is no limit on the number of log streams that you can create for a log group.

You must use the following guidelines when naming a log stream:
\itemize{
\item Log stream names must be unique within the log group.
\item Log stream names can be between 1 and 512 characters long.
\item The ':' (colon) and '*' (asterisk) characters are not allowed.
}
}
\examples{

}
