% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_destination}
\alias{delete_destination}
\title{Deletes the specified destination, and eventually disables all the subscription filters that publish to it}
\usage{
delete_destination(destinationName)
}
\arguments{
\item{destinationName}{The name of the destination.}
}
\description{
Deletes the specified destination, and eventually disables all the subscription filters that publish to it. This operation does not delete the physical resource encapsulated by the destination.
}
\examples{

}
