% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_log_group}
\alias{delete_log_group}
\title{Deletes the specified log group and permanently deletes all the archived log events associated with the log group}
\usage{
delete_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Deletes the specified log group and permanently deletes all the archived log events associated with the log group.
}
\section{Accepted Parameters}{
\preformatted{delete_log_group(
  logGroupName = "string"
)
}
}

