% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_log_stream}
\alias{delete_log_stream}
\title{Deletes the specified log stream and permanently deletes all the archived log events associated with the log stream}
\usage{
delete_log_stream(logGroupName, logStreamName)
}
\arguments{
\item{logGroupName}{The name of the log group.}

\item{logStreamName}{The name of the log stream.}
}
\description{
Deletes the specified log stream and permanently deletes all the archived log events associated with the log stream.
}
\examples{

}
