% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_resource_policy}
\alias{delete_resource_policy}
\title{Deletes a resource policy from this account}
\usage{
delete_resource_policy(
  policyName = "string"
)
}
\arguments{
\item{policyName}{The name of the policy to be revoked. This parameter is required.}
}
\description{
Deletes a resource policy from this account. This revokes the access of the identities in that policy to put log events to this account.
}
\examples{

}
