% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_retention_policy}
\alias{delete_retention_policy}
\title{Deletes the specified retention policy}
\usage{
delete_retention_policy(
  logGroupName = "string"
)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}
}
\description{
Deletes the specified retention policy.
}
\details{
Log events do not expire if they belong to log groups without a retention policy.
}
\examples{

}
