% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{describe_log_groups}
\alias{describe_log_groups}
\title{Lists the specified log groups}
\usage{
describe_log_groups(
  logGroupNamePrefix = "string",
  nextToken = "string",
  limit = 123
)
}
\arguments{
\item{logGroupNamePrefix}{The prefix to match.}

\item{nextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{limit}{The maximum number of items returned. If you don't specify a value, the default is up to 50 items.}
}
\description{
Lists the specified log groups. You can list all your log groups or filter the results by prefix. The results are ASCII-sorted by log group name.
}
\examples{

}
