% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{describe_subscription_filters}
\alias{describe_subscription_filters}
\title{Lists the subscription filters for the specified log group}
\usage{
describe_subscription_filters(logGroupName, filterNamePrefix = NULL,
  nextToken = NULL, limit = NULL)
}
\arguments{
\item{logGroupName}{The name of the log group.}

\item{filterNamePrefix}{The prefix to match. If you don't specify a value, no prefix filter is applied.}

\item{nextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{limit}{The maximum number of items returned. If you don't specify a value, the default is up to 50 items.}
}
\description{
Lists the subscription filters for the specified log group. You can list all the subscription filters or filter the results by prefix. The results are ASCII-sorted by filter name.
}
\examples{

}
