% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{list_tags_log_group}
\alias{list_tags_log_group}
\title{Lists the tags for the specified log group}
\usage{
list_tags_log_group(logGroupName)
}
\arguments{
\item{logGroupName}{The name of the log group.}
}
\description{
Lists the tags for the specified log group.
}
\examples{

}
