% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_package.R
\docType{package}
\name{paws.cloudwatchlogs-package}
\alias{paws.cloudwatchlogs}
\alias{paws.cloudwatchlogs-package}
\title{paws.cloudwatchlogs: Amazon CloudWatch Logs}
\description{
You can use Amazon CloudWatch Logs to monitor,
store, and access your log files from Amazon EC2 instances, AWS
CloudTrail, or other sources. You can then retrieve the associated log
data from CloudWatch Logs using the CloudWatch console, CloudWatch
Logs commands in the AWS CLI, CloudWatch Logs API, or CloudWatch Logs
SDK.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
