% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{put_destination_policy}
\alias{put_destination_policy}
\title{Creates or updates an access policy associated with an existing destination}
\usage{
put_destination_policy(destinationName, accessPolicy)
}
\arguments{
\item{destinationName}{A name for an existing destination.}

\item{accessPolicy}{An IAM policy document that authorizes cross-account users to deliver their log events to the associated destination.}
}
\description{
Creates or updates an access policy associated with an existing destination. An access policy is an \href{http://docs.aws.amazon.com/IAM/latest/UserGuide/policies_overview.html}{IAM policy document} that is used to authorize claims to register a subscription filter against a given destination.
}
\examples{

}
