% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{put_metric_filter}
\alias{put_metric_filter}
\title{Creates or updates a metric filter and associates it with the specified log group}
\usage{
put_metric_filter(logGroupName, filterName, filterPattern,
  metricTransformations)
}
\arguments{
\item{logGroupName}{The name of the log group.}

\item{filterName}{A name for the metric filter.}

\item{filterPattern}{A filter pattern for extracting metric data out of ingested log events.}

\item{metricTransformations}{A collection of information that defines how metric data gets emitted.}
}
\description{
Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents.
}
\details{
The maximum number of metric filters that can be associated with a log group is 100.
}
\examples{

}
