% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{put_retention_policy}
\alias{put_retention_policy}
\title{Sets the retention of the specified log group}
\usage{
put_retention_policy(
  logGroupName = "string",
  retentionInDays = 123
)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{retentionInDays}{[required]}
}
\description{
Sets the retention of the specified log group. A retention policy allows you to configure the number of days for which to retain log events in the specified log group.
}
\examples{

}
