% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{stop_query}
\alias{stop_query}
\title{Stops a CloudWatch Logs Insights query that is in progress}
\usage{
stop_query(
  queryId = "string"
)
}
\arguments{
\item{queryId}{[required] The ID number of the query to stop. If necessary, you can use \code{DescribeQueries} to find this ID number.}
}
\description{
Stops a CloudWatch Logs Insights query that is in progress. If the query has already ended, the operation returns an error indicating that the specified query is not running.
}
\examples{

}
