% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{tag_log_group}
\alias{tag_log_group}
\title{Adds or updates the specified tags for the specified log group}
\usage{
tag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{The name of the log group.}

\item{tags}{The key-value pairs to use for the tags.}
}
\description{
Adds or updates the specified tags for the specified log group.
}
\details{
To list the tags for a log group, use ListTagsLogGroup. To remove tags, use UntagLogGroup.

For more information about tags, see \href{http://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/log-group-tagging.html}{Tag Log Groups in Amazon CloudWatch Logs} in the \emph{Amazon CloudWatch Logs User Guide}.
}
\examples{

}
