% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{test_metric_filter}
\alias{test_metric_filter}
\title{Tests the filter pattern of a metric filter against a sample of log event messages}
\usage{
test_metric_filter(
  filterPattern = "string",
  logEventMessages = list(
    "string"
  )
)
}
\arguments{
\item{filterPattern}{[required]}

\item{logEventMessages}{[required] The log event messages to test.}
}
\description{
Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.
}
\examples{

}
