% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{untag_log_group}
\alias{untag_log_group}
\title{Removes the specified tags from the specified log group}
\usage{
untag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{tags}{[required] The tag keys. The corresponding tags are removed from the log group.}
}
\description{
Removes the specified tags from the specified log group.
}
\details{
To list the tags for a log group, use ListTagsLogGroup. To add tags, use UntagLogGroup.
}
\section{Accepted Parameters}{
\preformatted{untag_log_group(
  logGroupName = "string",
  tags = list(
    "string"
  )
)
}
}

