% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_metric_filter}
\alias{delete_metric_filter}
\title{Deletes the specified metric filter}
\usage{
delete_metric_filter(logGroupName, filterName)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{filterName}{[required] The name of the metric filter.}
}
\description{
Deletes the specified metric filter.
}
\section{Accepted Parameters}{
\preformatted{delete_metric_filter(
  logGroupName = "string",
  filterName = "string"
)
}
}

