% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{delete_subscription_filter}
\alias{delete_subscription_filter}
\title{Deletes the specified subscription filter}
\usage{
delete_subscription_filter(
  logGroupName = "string",
  filterName = "string"
)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{filterName}{[required] The name of the subscription filter.}
}
\description{
Deletes the specified subscription filter.
}
\examples{

}
