% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudwatchlogs_operations.R
\name{describe_resource_policies}
\alias{describe_resource_policies}
\title{Lists the resource policies in this account}
\usage{
describe_resource_policies(nextToken = NULL, limit = NULL)
}
\arguments{
\item{nextToken}{}

\item{limit}{The maximum number of resource policies to be displayed with one call of this API.}
}
\description{
Lists the resource policies in this account.
}
\section{Accepted Parameters}{
\preformatted{describe_resource_policies(
  nextToken = "string",
  limit = 123
)
}
}

